# MNT speed up test_kernel_pca_solvers_equivalence