@franslarsson 2021-11-25
> Small nitpicks but otherwise LGTM

Thanks for the review. I have committed the change.